Render single field for oneof
When using printer with `includingDefaultValueFields` all primitive
fields of `oneof` are rendered.

Then it's unclear which field was actualy set.

As I understand only one field from `oneof` should be renderd
or none if it was set to empty value.

I think this is a breaking change since I had to change tests in `JsonFormatSpec`.